deleted actions/reducers
